add: kanikoを起動する処理を書いた
