+++ .gitignore
@@ -1,4 +1,3 @@
 .env
-config.json
-build/
+build/config
 main
