+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    container_name: go-service
+    container_name: Eploy-service
     ports:
       - "8088:8088"
     tty: true
