+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/kouxi08/Eploy/handler"
-	
+
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 )
@@ -23,13 +23,14 @@ func server() {
 	e.Use(middleware.CORS())
 
 	//podのログを取得(クエリパラメータ,podName="ポッド名")
-	e.GET("/getpodlog",handler.GetPodLogHandler)
+	e.GET("/getpodlog", handler.GetPodLogHandler)
 
 	//レコード追加処理へ
 	e.POST("/", handler.CreateHandler)
-	
+
+	e.POST("/kaniko", handler.CreateKanikoHandler)
+
 	//レコード削除処理へ
 	e.PATCH("/", handler.DeleteHandler)
 	e.Logger.Fatal(e.Start(":8088"))
 }
-
