+++ pkg/kubernetes/new-client.go
@@ -1,0 +1,28 @@
+package kubernetes
+
+import (
+	"log"
+	"path/filepath"
+
+	// "errors"
+
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/tools/clientcmd"
+	"k8s.io/client-go/util/homedir"
+)
+
+func NewKubernetesClient() (*kubernetes.Clientset, error) {
+	kubeconfig := filepath.Join(homedir.HomeDir(), ".kube", "config")
+
+	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+	if err != nil {
+		log.Fatal(err)
+	}
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		log.Fatal(err)
+	}
+	return clientset, nil
+}
+
+func int32Ptr(i int32) *int32 { return &i }
