+++ pkg/resource_management.go
@@ -1,24 +1,30 @@
 package pkg
 
+import "github.com/kouxi08/Eploy/pkg/kubernetes"
+
 func CreateResources(siteName string, deploymentName string, serviceName string, ingressName string, hostName string) {
 	//deployment作成
-	CreateDeployment(siteName, deploymentName)
+	kubernetes.CreateDeployment(siteName, deploymentName)
 	//service作成
-	CreateService(siteName, serviceName)
+	kubernetes.CreateService(siteName, serviceName)
 	//ingress作成
-	CreateIngress(ingressName, hostName, serviceName)
+	kubernetes.CreateIngress(ingressName, hostName, serviceName)
+}
+
+func CreateKanikoResouces() {
+	kubernetes.CreateKaniko()
 }
 
 func DeleteResources(deploymentName string, serviceName string, ingressName string) {
 	//deployment削除
-	DeleteDeployment(deploymentName)
+	kubernetes.DeleteDeployment(deploymentName)
 	//service削除
-	DeleteService(serviceName)
+	kubernetes.DeleteService(serviceName)
 	//ingress削除
-	DeleteIngress(ingressName)
-}
-func GetLogPodResources(podName string) (message string,err error) {
-	message,err = GetPodLog(podName)
-	return
+	kubernetes.DeleteIngress(ingressName)
 }
 
+func GetLogPodResources(podName string) (message string, err error) {
+	message, err = kubernetes.GetPodLog(podName)
+	return
+}
